Rework the logic for suffixes and lookup paths
The definition of lookup paths cannot rely on the DIRS variable,
which contains only directories in the source tree, while we need
to find some files in the object tree.  They therefore need to be
initialised by the “module” module, as it examines the list of
products in the software package.

The list of suffixes is moved to the general initialisation file,
as it is a characteristic of the framework and must have been processed
before initialising lookup paths.