Do not use all of CFLAGS
It is wrong to pass all of CFLAGS to the mkdep program, as it
only understands the -I option for include directories.
Worse, the -p flag used to add profiling information to compiled objects
has a completely different meaning for mkdep, which it instructs
to build dependencies for programs rather than for objects.