BSD Owl Scripts v3.0.0-20141203
This pre-release shows improvements in the support of TeX and friends,
OCaml, scripts, adds support for C (with cross-compilation) and more.

Features:
- Support configuration profiles, so that we can easily switch
  from Debug to Release, for instance.
- The subhsell works even better than before.
- Support architectures, so that we can cross-compile our programs.
- Support C modules.
- Support OCaml modules.
- Support TeX modules, producing documents with index, bibliography,
  output to PDF, PostScript and DVI.
- Support METAPOST modules, producing pictures to use in TeX documents
  and as standalone documents.
- Support shell scripts and Python scripts.
- Support the M4 generic pre-processor.
- Support literate programming with noweb.
- Support parallel mode, allowing concurrent production of targets.

Our automated testsuite completed successfully under:
- Debian 3.16.7-2 (2014-11-06)
- Darwin 10.8.0
- FreeBSD 10.1-RELEASE
- CYGWIN_NT-5.1 1.7.32 (0.274/5/3)

Support of NetBSD and OpenBSD is delayed until the version of make
found in the base system is updated.

This is alpha software and its distribution is intended for testing
purposes only.