Copy the files instead of linking them
When using hard-links, the new file names and the old file names
refer to the same inode.  They therefore share their owner,
group and permissions, but we want to alter the owner, group
and permissions of the data referred to by the new file names
independently from the owner, group and permissions of the data
referred to by the old file names.

We therefore remove the -l flag from the cpio command.