Fix definition of MAKEOBJDIR
There is one subtle point to consider when crafting a MAKEOBJDIR variable
using the CONFIGURATION and ARCHITECTURE variable. The point is that
the .OBJDIR directory is made before everything else, which means that
variables CONFIGURATIONDIR and ARCHITECTUREDIR deduced from the two
others are *not* available at the time where .OBJDIR is computed.
This leads to a odd and hard to diagnose behaviour.

Instead of using CONFIGURATIONDIR and ARCHITECTUREDIR one can use

  ${CONFIGURATION:C@^.@/&@} and ${ARCHITECTURE:C@^.@/&@}

in the definition of MAKEOBJDIR.